move dependencies of erisdb/ on evm, state and account to manager/erismint
